# openapi.yml
openapi: 3.1.2
info:
  title: "短网址"
  description: "短网址接口"
  contact:
    name: "Jetsung Chan"
    url: "https://github.com/jetsung"
    email: "i@jetsung.com"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0
servers:
  - url: http://127.0.0.1:8080
    description: Version 1
tags:
  - name: shorten
    description: 短址
  - name: history
    description: 历史记录
  - name: account
    description: 账号
paths:
  /api/account/login:
    post:
      tags:
        - account
      description: 登录接口
      operationId: login
      requestBody:
        description: 登录系统
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginParams"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      x-codegen-request-body-name: body
    x-swagger-router-controller: api

  /api/account/logout:
    post:
      description: 退出登录接口
      operationId: logout
      tags:
        - account
      responses:
        "204":
          description: Success
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    x-swagger-router-controller: api

  /api/users/current:
    get:
      tags:
        - account
      description: 获取当前的用户
      operationId: currentUser
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    x-swagger-router-controller: api

  /api/shortens:
    post:
      tags:
        - shorten
      summary: "添加短网址"
      description: "添加一个新的短网址"
      operationId: "addShorten"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Shorten"
      responses:
        "201":
          description: "短网址创建成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenResponse"
        "400":
          description: "请求错误"
        "409":
          description: "短网址已存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      tags:
        - shorten
      summary: "获取所有短址信息"
      description: "获取所有短址信息"
      operationId: "getShortens"
      parameters:
        - name: page
          in: query
          description: "页码"
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: "每页条数"
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          description: "排序字段"
          required: false
          schema:
            type: string
            default: "created_at"
            enum:
              - id
              - short_code
              - created_at
              - updated_at
        - name: order
          in: query
          description: "排序方向"
          required: false
          schema:
            type: string
            default: "desc"
            enum:
              - asc
              - desc
        - name: status
          in: query
          description: "状态"
          required: false
          schema:
            type: integer
            default: 0
            enum:
              - 0
              - 1
              - 2
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ShortenResponse"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: "请求错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - shorten
      summary: "删除短网址列表"
      description: "批量删除短网址"
      operationId: "deleteShortenBatch"
      parameters:
        - name: ids
          in: query
          description: "ID 列表，多个 ID 用逗号分隔"
          required: true
          schema:
            type: string
            pattern: ^\d+(,\d+)*$
      responses:
        "204":
          description: "操作成功"
        "400":
          description: "请求错误"
        "404":
          description: "短网址不存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/shortens/{short_code}:
    get:
      tags:
        - shorten
      summary: "获取短网址信息"
      description: "通过短码获取短网址信息"
      operationId: "getShorten"
      parameters:
        - name: short_code
          in: path
          description: "短码"
          required: true
          schema:
            type: string
            maxLength: 16
            pattern: "^[a-zA-Z0-9]+$"
            example: "aBc123"
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenResponse"
        "400":
          description: "请求错误"
        "404":
          description: "短网址不存在"
        "500":
          description: "操作失败"
        "401":
          description: "未授权"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - shorten
      summary: "更新短网址"
      description: "更新一个短网址"
      operationId: "updateShorten"
      parameters:
        - name: short_code
          in: path
          description: "短码"
          required: true
          schema:
            type: string
            maxLength: 16
            pattern: "^[a-zA-Z0-9]+$"
            example: "aBc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShortenUpdate"
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortenResponse"
        "400":
          description: "请求错误"
        "404":
          description: "短码不存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - shorten
      summary: "删除短网址"
      description: "删除一个短网址"
      operationId: "deleteShorten"
      parameters:
        - name: short_code
          in: path
          description: "短码"
          required: true
          schema:
            type: string
            maxLength: 16
            pattern: "^[a-zA-Z0-9]+$"
            example: "aBc123"
      responses:
        "204":
          description: "操作成功"
        "400":
          description: "请求错误"
        "404":
          description: "短网址不存在"
        "500":
          description: "操作失败"
        "401":
          description: "未授权"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/histories:
    get:
      tags:
        - history
      summary: "获取访问历史记录"
      description: "获取所有访问日志信息"
      operationId: "getHistories"
      parameters:
        - name: page
          in: query
          description: "页码"
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: "每页条数"
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          description: "排序字段"
          required: false
          schema:
            type: string
            default: "accessed_at"
            enum:
              - id
              - accessed_at
              - created_at
        - name: order
          in: query
          description: "排序方向"
          required: false
          schema:
            type: string
            default: "desc"
            enum:
              - asc
              - desc
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/HistoryResponse"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: "请求错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags:
        - history
      summary: "删除访问日志"
      description: "批量删除访问日志"
      operationId: "deleteHistories"
      parameters:
        - name: ids
          in: query
          description: "ID 列表，多个 ID 用逗号分隔"
          required: true
          schema:
            type: string
            pattern: ^\d+(,\d+)*$
      responses:
        "204":
          description: "操作成功"
        "400":
          description: "请求错误"
        "404":
          description: "日志不存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    PageMeta:
      type: object
      required:
        - page
        - per_page
        - count
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
          description: 当前页码
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          description: 每页数量
          example: 10
        count:
          type: integer
          minimum: 0
          description: 当前页条目数
          example: 10
        total:
          type: integer
          minimum: 0
          description: 总条目数
          example: 100
        total_pages:
          type: integer
          minimum: 0
          description: 总页数
          example: 10

    LoginParams:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          format: password
          description: 密码
        auto_login:
          type: boolean
          description: 是否自动登录
          default: false

    LoginResult:
      type: object
      properties:
        token:
          type: string
          description: 登录成功后返回的 token
        error_code:
          type: string
          description: 错误码
        error_message:
          type: string
          description: 错误信息

    CurrentUser:
      type: object
      properties:
        name:
          type: string
          description: 用户名

    ErrorResponse:
      required:
        - error_code
      type: object
      properties:
        error_code:
          type: string
          description: 业务约定的错误码
        error_message:
          type: string
          description: 错误信息

    Shorten:
      type: object
      required:
        - original_url
        - short_code
      properties:
        original_url:
          type: string
          format: uri
          description: "原始长网址"
          example: "https://example.com/very/long/url"
        short_code:
          type: string
          description: "自定义短码"
          maxLength: 16
          minLength: 3
          pattern: "^[a-zA-Z0-9]+$"
          example: "abc123"
        description:
          type: string
          description: "短链描述"
          maxLength: 255

    ShortenUpdate:
      type: object
      required:
        - original_url
      properties:
        original_url:
          type: string
          format: uri
          description: "原始长网址"
        description:
          type: string
          description: "短链描述"

    ShortenResponse:
      type: object
      properties:
        id:
          type: integer
          description: "数据库 ID"
        short_code:
          type: string
          description: "短码"
          maxLength: 16
          pattern: "^[a-zA-Z0-9]+$"
          example: "abc123"
        short_url:
          type: string
          format: uri
          description: "完整短网址"
          example: "https://example.com/s/abc123"
        original_url:
          type: string
          format: uri
          description: "原始长网址"
        description:
          type: string
          description: "短链描述"
        status:
          type: integer
          description: "状态：0=启用, 1=禁用, 2=未知"
          enum: [0, 1, 2]
          default: 0
        created_at:
          type: string
          format: date-time
          description: "创建时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: "更新时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"

    HistoryResponse:
      type: object
      required:
        - id
        - url_id
        - short_code
        - ip_address
        - user_agent
        - accessed_at
        - created_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: 访问记录 ID
        url_id:
          type: integer
          format: int32
          description: 短链接 ID
        short_code:
          type: string
          pattern: ^[a-zA-Z0-9]{5,16}$
          example: "xY7z9"
          description: 短码
        ip_address:
          type: string
          format: ipv4
          example: "192.168.0.1"
          description: 访问者 IP
        user_agent:
          type: string
          minLength: 10
          example: "Mozilla/5.0 (Windows NT 10.0)"
          description: 浏览器 UA
        referer:
          anyOf:
            - type: string
              format: uri
            - type: "null"
          example: "https://google.com"
          description: 来源页面
        country:
          type: string
          example: "中国"
        region:
          type: string
          example: "华南"
        province:
          type: string
          example: "广东省"
        city:
          type: string
          example: "深圳市"
        isp:
          type: string
          example: "中国电信"
        device_type:
          type: string
          enum: [pc, mobile, tablet]
        os:
          type: string
          enum: [Windows, MacOS, Linux, Android, iOS]
        browser:
          type: string
          enum: [Chrome, Firefox, Safari, Edge, IE]
        accessed_at:
          type: string
          format: date-time
          description: "访问时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"
        created_at:
          type: string
          format: date-time
          description: "记录创建时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"